improve: move custom hook useTheme in separate file
